Initial commit
Added repo setup files (licenses, readme, gitignore, etc) as well
as some initial scripts.

The scripts are not in the format that I want to structure this
directory. For example, there is no scripts directory with executable
scripts. And the weather_api_interaction is a large chunk of code still.